added pilot image and deleted old game class
